# docs/swagger.yaml
basePath: /
definitions:
...
  handler.CreateGroupRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  handler.UpdateGroupRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: REST API untuk personal anime library.
  title: Libra - Anime Library API
  version: "1.0"
paths:
...
  /me/groups:
    get:
      description: Get all anime groups belonging to the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Groups retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user's groups
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Create a new anime group for the authenticated user.
      parameters:
      - description: Create group request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Group created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a group
      tags:
      - Groups
  /me/groups/{id}:
    delete:
      description: Delete a group belonging to the authenticated user.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Group deleted successfully
        "400":
          description: Invalid group ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a group
      tags:
      - Groups
    get:
      description: Get a specific group belonging to the authenticated user.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Group retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid group ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get group detail
      tags:
      - Groups
    patch:
      consumes:
      - application/json
      description: Update the name or description of a group belonging to the authenticated
        user.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update group request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Group updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid token
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a group
      tags:
      - Groups
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"